did c3prj2

+++ c3prj1_deck/deck.c
@@ -3,14 +3,14 @@
 #include <assert.h>
 #include "deck.h"
 void print_hand(deck_t * hand){
-  for (int i=0;i<(*hand).n_cards;i++){
+  for (size_t i=0;i<(*hand).n_cards;i++){
     print_card(*(*hand).cards[i]);
     printf(" ");
   }
 }
 
 int deck_contains(deck_t * d, card_t c) {
-  for(int i=0;i<(*d).n_cards;i++){
+  for(size_t i=0;i<(*d).n_cards;i++){
     if((*(*d).cards[i]).value==c.value && (*(*d).cards[i]).suit==c.suit){
       return 1;
     }
@@ -19,7 +19,7 @@ int deck_contains(deck_t * d, card_t c) {
 }
 
 void shuffle(deck_t * d){
-  for (int i=0; i<(*d).n_cards/2;i++){
+  for (size_t i=0; i<(*d).n_cards/2;i++){
     size_t a = rand()%(*d).n_cards;
     size_t b = rand()%(*d).n_cards;
     card_t* tmp = (*d).cards[a];
